sources/iotapi/requirements.txt
asciimatics==1.12.0
attrs==20.3.0
aws-requests-auth==0.4.3
AWSIoTPythonSDK==1.4.9
bcrypt==3.2.0
boto3==1.17.20
botocore==1.20.20
certifi==2020.11.8
cffi==1.14.6
chardet==3.0.4
click==7.1.2
colorama==0.4.4
commonmark==0.9.1
configparser==5.2.0
coverage==5.3
cryptography==3.4.7
decorator==4.4.2
dnslib==0.9.14
dnspython==1.16.0
docker-py==1.10.6
docker-pycreds==0.4.0
docopt==0.6.2
Faker==4.16.0
future==0.18.2
idna==2.10
iniconfig==1.1.1
invoke==1.5.0
Jinja2==2.11.2
jmespath==0.10.0
keyring==22.3.0
liccheck==0.6.5
MarkupSafe==1.1.1
packaging==20.4
paramiko==2.7.2
Pillow==8.3.1
pip-licenses==3.5.4
pluggy==0.13.1
pony==0.7.14
prettytable==2.0.0
prompt-toolkit==3.0.17
psycopg2==2.9.3
PTable==0.9.2
py==1.9.0
pyaes==1.6.1
pycparser==2.20
pyfiglet==0.8.post1
Pygments==2.7.2
PyNaCl==1.4.0
pynamodb==4.3.3
pyparsing==2.4.7
pypng==0.0.20
PyQRCodeNG==1.3.6
pytest==6.1.2
python-dateutil==2.8.1
python-dotenv==0.19.0
questionary==1.9.0
requests==2.25.0
requests-aws4auth==1.0.1
rich==9.2.0
s3transfer==0.3.3
semantic-version==2.9.0
six==1.15.0
sqlitedict==1.7.0
sshtunnel==0.3.1
text-unidecode==1.3
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.2
validators==0.18.2
wcwidth==0.2.5
websocket-client==0.57.0
